Made improvement to address PR comments
