+++ modules/nft/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 serde = { version = "1.0.124", optional = true }
 codec = { package = "parity-scale-codec", version = "2.3.1", default-features = false, features = ["max-encoded-len"] }
 scale-info = { version = "1.0", default-features = false, features = ["derive"] }
-enumflags2 = { version = "0.6.3" }
 
 frame-benchmarking = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.15", default-features = false, optional = true}
 frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.15", default-features = false }
@@ -36,7 +35,6 @@ default = ["std"]
 std = [
 	"serde",
 	"codec/std",
-	"enumflags2/serde",
 	"scale-info/std",
 	"sp-std/std",
 	"sp-runtime/std",
@@ -46,7 +44,6 @@ std = [
 	"primitives/std",
 	"orml-traits/std",
 	"orml-nft/std",
-	"enumflags2/serde",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking/runtime-benchmarks",
