+++ modules/nft/src/benchmarking.rs
@@ -26,6 +26,7 @@ use frame_benchmarking::{account, benchmarks};
 use frame_support::{dispatch::DispatchErrorWithPostInfo, traits::Get, weights::DispatchClass};
 use frame_system::RawOrigin;
 use sp_runtime::traits::{AccountIdConversion, StaticLookup, UniqueSaturatedInto};
+use sp_std::collections::btree_map::BTreeMap;
 
 pub use crate::*;
 use primitives::Balance;
